Defined callback function compute_schema/2.
Takes in a list of input schemas as well as the wrapped
(implementation-specific) part of the plan node state, and produces the
output schema of the plan node.

The state is necessary in cases where the changes made to the schema
depend on arguments given to init/1, e.g. the project specification.