Prepares schema for plan node on its own.
Changed {listen, PlanNode :: pid()} option to
{listen, PlanNodes :: [pid()]}. The plan node will then add itself as
a listener to all of these plan nodes.

The plan node then constructs the schema for this node using the newly
defined compute_schema/2 function.

Amended the contract of the add_listener/2 to show that it returns a
schema, rather than simply the atom 'ok'.