Created ydb_planner module.
Defined initial structure for how queries are processed. The
intermediate nodes are able to listen to either (1) an input stream,
or (2) another intermediate node.

Each intermediate node in a query is identified by a tuple consisting
its type, and the number of a similar type before it, e.g. {select, 0}.
Not completely sure how this will work when attempting to introduce
joins to the planner and need to keep the child identifiers unique.
On second thought, should be fine so long as pass the resulting history
dict from the first call to the second.

Attempted to define representation of select node.