Refactored how plan node adds itself as a listener.
Added support for listen option to contain a list of functions that
resolve to a pid. This is necessary in order to integrate with the
supervisor hierarchy because the pids of the previous plan nodes are
not necessarily constant.

The function acts as a closure to retrieve the pid of an intermediate
node with a specified name from the query.

Because the plan node process would reach deadlock if attempted to
call this function, since it calls to supervisor:which_children/1.
Thus, the plan node defers the act of computing the schema until after
it has finished initializing (when its supervisor would actually be
able to respond). Note that because the message is sent to itself
during init/1, it is guaranteed to be the first message in its mailbox
post-initialization, so no race conditions exist of say the next plan
node requesting the schema from the current one.