Began rewriting ydb_planner module.
Inverted the nesting order of how queries are defined. Rather than
going from the input stream to its listener and so on, the nested tuple
is node in the plan that the one currently processed listens to.

As written, this function make/1 is not tail-recursive, but could be
done so using continuation-passing style.