Debugged and finished join node algorithm.
Had made the mistake of using the parity to determine whether or not
a output diff should be sent to the subscribers. Whereas, the only
consideration that needed to be made was whether during the backward
join process, if the table had just been created or not.

Also, looks like a couple off-by-one errors with the history size.