Implemented logic for processing received diffs.
Extracts the PLUS and MINUS tuples separately and calls do_insert/2
and do_delete/2 on them.

Fixed an issue with do_insert/2 where was not catching the badarg
exception correctly.

Fixed an issue with get_relation/1 where was not constructing a multi-
set based on the count of tuple in the synopsis.