Created generic publisher process.
The `ydb_gen_pub` module is responsible for the un/subscription
procedure for processes in a topology. Each node has a schema that is
then passed to any subscribers so that they may dynamically compute
their own schema.

This functionality encompasses much of the existing code in
`ydb_plan_node`.

The intent is for the bolt and spout implementation to extend this
module, and define further callbacks as necessary. Note also that only
the `ydb_gen_bolt` module needs to handle the interactions with the
`ydb_ets_mgr` module.