Added support for creating managed ETS tables.
The heir of the table is set as the manager, so that in case the
receiving process dies, the table under ownership of the manager.

Currently the manager does _not_ support state recovery; meaning that
requests made by a restarted process for a table will **always** return
a new ETS table right now.