Defined interface function new_synopsis/1.
Calls to ydb_ets_mgr:new_ets/3 with an appropriate default
configuration, i.e. name and set of options for the ETS table. Note
that the tables are either given a name 'synopsis' or 'synopsis1',
'synopsis2', ...

The result is yielded back to the process via the receive_synopsis/2
callback function.